# collections/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: 1.0.0
  title: 'Swagger JsonApi'
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: localhost
schemes:
  - https
  - http
paths:
  /tickets/:
    get:
      tags:
        - tickets
      summary: 'Ticket List'
      operationId: listTicket
      produces:
        - application/json
      parameters: {  }
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: '/tickets?page[number]=1&page[size]=100'
                  first:
                    type: string
                    example: '/tickets?page[number]=1&page[size]=100'
                  last:
                    type: string
                    example: '/tickets?page[number]=1&page[size]=100'
                  prev:
                    type: string
                    example: 'null'
                  next:
                    type: string
                    example: 'null'
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                      example: 12
                    type:
                      type: string
                      example: tickets
                    attributes:
                      $ref: '#/definitions/Ticket'
                    relationships:
                      properties: {  }
    post:
      tags:
        - tickets
      summary: 'Add Ticket'
      operationId: addTicket
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: addTicket
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    example: tickets
                  attributes:
                    $ref: '#/definitions/Ticket'
                  relationships:
                    properties: {  }
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /tickets
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: tickets
                  attributes:
                    $ref: '#/definitions/Ticket'
                  relationships:
                    properties: {  }
  '/tickets/{Ticket_id}':
    patch:
      tags:
        - tickets
      summary: 'Edit Ticket'
      operationId: editTicket
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: editTicket
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: tickets
                  attributes:
                    $ref: '#/definitions/Ticket'
                  relationships:
                    properties: {  }
        -
          name: Ticket_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /tickets
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: tickets
                  attributes:
                    $ref: '#/definitions/Ticket'
                  relationships:
                    properties: {  }
    delete:
      tags:
        - tickets
      summary: 'Delete Ticket'
      operationId: deleteTicket
      produces:
        - application/json
      parameters:
        -
          name: Ticket_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /tickets
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: tickets
                  attributes:
                    $ref: '#/definitions/Ticket'
                  relationships:
                    properties: {  }
    get:
      tags:
        - tickets
      summary: 'Get Ticket'
      operationId: viewTicket
      produces:
        - application/json
      parameters:
        -
          name: Ticket_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /tickets
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: tickets
                  attributes:
                    $ref: '#/definitions/Ticket'
                  relationships:
                    properties: {  }
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Ticket:
    type: object
    properties:
      ticketId:
        type: integer
        format: int32
      ticketCode:
        type: string
      scanTimestamp:
        type: integer
        format: int32
      device:
        type: string
externalDocs:
  description: 'Find out more about Swagger'
  url: 'http://swagger.io'